+++ app/src/main/java/com/dikin/shoppingapp/entities/Category.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "categories")
 data class Category(
     @PrimaryKey(autoGenerate = true)
-    val categoryId: Int = 0,
+    val categoryId: Long = 0,
     val categoryName: String
 )
