+++ app/src/main/java/com/dikin/shoppingapp/models/CartItemWithProduct.kt
@@ -1,9 +1,9 @@
 package com.dikin.shoppingapp.models
 
 data class CartItemWithProduct(
-    val cartItemId: Int,
-    val cartId: Int,
-    val productId: Int,
+    val cartItemId: Long,
+    val cartId: Long,
+    val productId: Long,
     val quantity: Int,
     val productName: String,
     val productDescription: String,
