+++ app/src/main/java/com/dikin/shoppingapp/viewmodels/ReviewViewModel.kt
@@ -21,7 +21,7 @@ class ReviewViewModel(application: Application) : AndroidViewModel(application) 
         all = repository.all.asLiveData()
     }
 
-    fun getById(id: Int, callback: (Review?) -> Unit) = viewModelScope.launch {
+    fun getById(id: Long, callback: (Review?) -> Unit) = viewModelScope.launch {
         callback(repository.getById(id))
     }
 
@@ -37,7 +37,7 @@ class ReviewViewModel(application: Application) : AndroidViewModel(application) 
         repository.delete(review)
     }
 
-    fun deleteById(id: Int) = viewModelScope.launch {
+    fun deleteById(id: Long) = viewModelScope.launch {
         repository.deleteById(id)
     }
 }
